Extract dictionaries & fetchDictionary
