+++ packages/scrabble-solver/src/service-worker/routeSolveRequests.ts
@@ -1,9 +1,10 @@
-import { Trie } from '@kamilmielnik/trie';
 import { getConfig } from '@scrabble-solver/configs';
 import { BLANK } from '@scrabble-solver/constants';
 import { solve } from '@scrabble-solver/solver';
 import { Board, Locale, Tile } from '@scrabble-solver/types';
 import { registerRoute } from 'workbox-routing';
+
+import { getDictionary } from './dictionaries';
 
 const headers = {
   'Content-Type': 'application/json; charset=utf-8',
@@ -14,7 +15,7 @@ const routeSolveRequests = () => {
     ({ url }) => url.pathname === '/api/solve',
     async ({ request }) => {
       const { board, characters, configId, locale } = await request.clone().json();
-      const trie = getTrie(locale);
+      const trie = getDictionary(locale);
 
       if (!trie) {
         return fetch(request);
@@ -28,44 +29,6 @@ const routeSolveRequests = () => {
     },
     'POST',
   );
-};
-
-const cache: Partial<Record<Locale, Trie | Promise<Trie> | undefined>> = {};
-
-const getTrie = (locale: Locale): Trie | undefined => {
-  const cached = cache[locale];
-
-  if (cached) {
-    if (cached instanceof Promise) {
-      return undefined;
-    }
-
-    return cached;
-  }
-
-  const promise = fetchTrie(locale);
-  cache[locale] = promise;
-  promise
-    .then((trie) => {
-      cache[locale] = trie;
-    })
-    .catch(() => {
-      cache[locale] = undefined;
-    });
-
-  return undefined;
-};
-
-const fetchTrie = async (locale: Locale): Promise<Trie> => {
-  const response = await fetch(`/api/dictionary/${locale}`);
-
-  if (!response.ok) {
-    throw new Error(`HTTP ${response.status}: ${response.statusText}`);
-  }
-
-  const serialized = await response.text();
-  const trie = Trie.deserialize(serialized);
-  return trie;
 };
 
 export default routeSolveRequests;
